+++ android/apollo/src/main/java/io/muun/apollo/data/os/BackgroundExecutionMetricsProvider.kt
@@ -28,6 +28,8 @@ class BackgroundExecutionMetricsProvider @Inject constructor(
     private val systemCapabilitiesProvider: SystemCapabilitiesProvider,
     private val dateTimeZoneProvider: DateTimeZoneProvider,
     private val localeInfoProvider: LocaleInfoProvider,
+    private val trafficStatsInfoProvider: TrafficStatsInfoProvider,
+    private val nfcProvider: NfcProvider,
 ) {
 
     private val powerManager: PowerManager by lazy {
@@ -72,7 +74,24 @@ class BackgroundExecutionMetricsProvider @Inject constructor(
             connectivityInfoProvider.proxySocks,
             dateTimeZoneProvider.autoDateTime,
             dateTimeZoneProvider.autoTimeZone,
-            dateTimeZoneProvider.timeZoneId
+            dateTimeZoneProvider.timeZoneId,
+            localeInfoProvider.dateFormat,
+            localeInfoProvider.regionCode,
+            dateTimeZoneProvider.calendarIdentifier,
+            trafficStatsInfoProvider.androidMobileRxTraffic,
+            telephonyInfoProvider.simOperatorId,
+            telephonyInfoProvider.simOperatorName,
+            telephonyInfoProvider.mobileNetworkId,
+            telephonyInfoProvider.mobileNetworkName,
+            telephonyInfoProvider.mobileRoaming,
+            telephonyInfoProvider.mobileDataStatus,
+            telephonyInfoProvider.mobileRadioType,
+            telephonyInfoProvider.mobileDataActivity,
+            connectivityInfoProvider.networkLink,
+            nfcProvider.hasNfcFeature(),
+            nfcProvider.hasNfcAdapter,
+            nfcProvider.isNfcEnabled,
+            nfcProvider.getNfcAntennaPosition().map { "${it.first};${it.second}" }.toTypedArray()
         )
 
     @Suppress("ArrayInDataClass")
@@ -114,7 +133,24 @@ class BackgroundExecutionMetricsProvider @Inject constructor(
         private val proxySocks: String,
         private val autoDateTime: Int,
         private val autoTimeZone: Int,
-        private val timeZoneId: String
+        private val timeZoneId: String,
+        private val androidDateFormat: String,
+        private val regionCode: String,
+        private val androidCalendarIdentifier: String,
+        private val androidMobileRxTraffic: Long,
+        private val androidSimOperatorId: String,
+        private val androidSimOperatorName: String,
+        private val androidMobileOperatorId: String,
+        private val mobileOperatorName: String,
+        private val androidMobileRoaming: Boolean,
+        private val androidMobileDataStatus: Int,
+        private val androidMobileRadioType: Int,
+        private val androidMobileDataActivity: Int,
+        private val androidNetworkLink: ConnectivityInfoProvider.NetworkLink?,
+        private val androidHasNfcFeature: Boolean,
+        private val androidHasNfcAdapter: Boolean,
+        private val androidNfcEnabled: Boolean,
+        private val androidNfcAntennaPositions: Array<String>
     )
 
     /**