+++ android/apollo/src/main/java/io/muun/apollo/domain/model/report/CrashReport.kt
@@ -2,6 +2,15 @@ package io.muun.apollo.domain.model.report
 
 import android.util.Log
 import java.io.Serializable
+import kotlin.math.min
+
+/**
+ * This acts as a safe-guard when sending errors in our own email reports.
+ * Too many data can cause crashes (e.g TransactionTooLargeException in when trying use emailIntent
+ * for email reports).
+ */
+private const val STACK_TRACE_LIMIT_FOR_EMAIL_REPORTS = 10_000
+private const val STACK_TRACE_LIMIT_FOR_ANALYTICS = 500
 
 data class CrashReport(
     val tag: String,
@@ -11,8 +20,8 @@ data class CrashReport(
     val metadata: MutableMap<String, Serializable>,
 ) {
 
-    fun print() =
-        "Tag:$tag\nMessage:$message\nError:${printError()}\nMetadata:{\n\n${printMetadata()}}"
+    fun print(abridged: Boolean) =
+        "Message:$message\nError:${printError(abridged)}\nMetadata:{\n\n${printMetadata()}}"
 
     fun printMetadata(): String {
         val builder = StringBuilder()
@@ -22,8 +31,23 @@ data class CrashReport(
         return builder.toString()
     }
 
-    fun printError(): String =
-        Log.getStackTraceString(error)
+    private fun printError(abridged: Boolean = false): String {
+        val stackTraceString = Log.getStackTraceString(error)
+        return if (abridged) {
+            val stackTraceCapSizeInChars =
+                min(STACK_TRACE_LIMIT_FOR_EMAIL_REPORTS, stackTraceString.length)
+            return "${stackTraceString.substring(0, stackTraceCapSizeInChars)}\nEXCEEDED MAX LENGTH"
+        } else {
+            stackTraceString
+        }
+    }
+
+    fun printErrorForAnalytics(): String {
+        val stackTraceString = Log.getStackTraceString(error)
+        val stackTraceCapSizeInChars = min(STACK_TRACE_LIMIT_FOR_ANALYTICS, stackTraceString.length)
+        return stackTraceString.substring(0, stackTraceCapSizeInChars)
+            .replace("\n", " ")
+    }
 
     fun getTrackingTitle(): String =
         error.javaClass.simpleName + ":" + error.localizedMessage?.replace("\n", " ")